Quantum Gamemode - Garry's Mod
This is a scripted custom gamemode for the game Garry's Mod. It utilizes
the source engine and by using LUA script we can manipulate the provided
assets to create a custom game. This is a simple RPG based game that
works by placing this in your gamemode folder in garry's mod.